engine: Performance improvement in clean up compensation data
If the command is running with out compensation context no need to clean it.
In memory check is much more faster than call to db

+++ backend/manager/modules/bll/src/main/java/org/ovirt/engine/core/bll/CommandBase.java
@@ -415,7 +415,9 @@ public abstract class CommandBase<T extends VdcActionParametersBase> extends Aud
      * Delete the compensation data, so that we don't accidentaly try to compensate it at a later time.
      */
     private void cleanUpCompensationData() {
-        getBusinessEntitySnapshotDAO().removeAllForCommandId(commandId);
+        if (!(getCompensationContext() instanceof NoOpCompensationContext)) {
+            getBusinessEntitySnapshotDAO().removeAllForCommandId(commandId);
+        }
     }
 
     public VdcReturnValueBase endAction() {
